--- src/Symfony/Bundle/SecurityBundle/Resources/config/form.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services https://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <services>
-        <defaults public="false" />
-
-        <!-- SecurityPasswordType -->
-        <service id="form.type.security_password" class="Symfony\Bundle\SecurityBundle\Form\SecurityPasswordType">
-            <argument type="service" id="security.user_password_encoder"/>
-            <tag name="form.type" alias="security_password" />
-        </service>
-    </services>
-</container>
